+++ gcs/CHANGES.md
@@ -130,6 +130,8 @@
 
 1.  Use default logging backend for Google Flogger instead of Slf4j.
 
+1.  Add `FsBenchmark` tool for benchmarking HCFS.
+
 ### 2.1.1 - 2020-03-11
 
 1.  Add upload cache to support high-level retries of failed uploads. Cache size